# docs/swagger/swagger.yaml
basePath: /v0
consumes:
- application/json
definitions:
  github_com_rabbytesoftware_quiver_core_internal_api_libs.ErrResponse:
    properties:
      error:
        type: string
      namespace:
        type: string
      success:
        type: boolean
    type: object
...
  github_com_rabbytesoftware_quiver_core_internal_api_libs.QueryResponse:
    properties:
      data: {}
      success:
        type: boolean
    type: object
...
  github_com_rabbytesoftware_quiver_core_internal_api_v0_dto.ConfigDTO:
    properties:
      configured:
        $ref: '#/definitions/github_com_rabbytesoftware_quiver_core_internal_app_usecases.Config'
      corrected:
        items:
          $ref: '#/definitions/github_com_rabbytesoftware_quiver_core_internal_api_v0_dto.ConfigRejectionDTO'
        type: array
      defaults:
        $ref: '#/definitions/github_com_rabbytesoftware_quiver_core_internal_app_usecases.Config'
      restart_required:
        items:
          type: string
        type: array
      running:
        $ref: '#/definitions/github_com_rabbytesoftware_quiver_core_internal_api_v0_dto.runningConfig'
    type: object
  github_com_rabbytesoftware_quiver_core_internal_api_v0_dto.ConfigPatchResultDTO:
    properties:
      applied:
        items:
          type: string
        type: array
      rejected:
        items:
          $ref: '#/definitions/github_com_rabbytesoftware_quiver_core_internal_api_v0_dto.ConfigRejectionDTO'
        type: array
    type: object
  github_com_rabbytesoftware_quiver_core_internal_api_v0_dto.ConfigRejectionDTO:
    properties:
      key:
        type: string
      message:
        type: string
    type: object
...
  github_com_rabbytesoftware_quiver_core_internal_api_v0_dto.PathStatusDTO:
    properties:
      bin_dir:
        type: string
      configured:
        type: boolean
      files:
        items:
          type: string
        type: array
      on_path:
        type: boolean
    type: object
...
  github_com_rabbytesoftware_quiver_core_internal_api_v0_dto.runningConfig:
    properties:
      api:
        $ref: '#/definitions/github_com_rabbytesoftware_quiver_core_internal_core_config.API'
      arrows:
        $ref: '#/definitions/github_com_rabbytesoftware_quiver_core_internal_core_config.Arrows'
      auth:
        $ref: '#/definitions/github_com_rabbytesoftware_quiver_core_internal_core_config.Auth'
      logger:
        $ref: '#/definitions/github_com_rabbytesoftware_quiver_core_internal_core_config.Logger'
      manifold:
        $ref: '#/definitions/github_com_rabbytesoftware_quiver_core_internal_core_config.Manifold'
      netbridge:
        $ref: '#/definitions/github_com_rabbytesoftware_quiver_core_internal_core_config.Netbridge'
      search:
        $ref: '#/definitions/github_com_rabbytesoftware_quiver_core_internal_core_config.Search'
      vault:
        $ref: '#/definitions/github_com_rabbytesoftware_quiver_core_internal_core_config.Vault'
    type: object
...
  github_com_rabbytesoftware_quiver_core_internal_app_usecases.Config:
    properties:
      api:
        $ref: '#/definitions/github_com_rabbytesoftware_quiver_core_internal_core_config.API'
      arrows:
        $ref: '#/definitions/github_com_rabbytesoftware_quiver_core_internal_core_config.Arrows'
      auth:
        $ref: '#/definitions/github_com_rabbytesoftware_quiver_core_internal_core_config.Auth'
      logger:
        $ref: '#/definitions/github_com_rabbytesoftware_quiver_core_internal_core_config.Logger'
      manifold:
        $ref: '#/definitions/github_com_rabbytesoftware_quiver_core_internal_core_config.Manifold'
      netbridge:
        $ref: '#/definitions/github_com_rabbytesoftware_quiver_core_internal_core_config.Netbridge'
      search:
        $ref: '#/definitions/github_com_rabbytesoftware_quiver_core_internal_core_config.Search'
      vault:
        $ref: '#/definitions/github_com_rabbytesoftware_quiver_core_internal_core_config.Vault'
    type: object
  github_com_rabbytesoftware_quiver_core_internal_core_config.API:
    properties:
      host:
        type: string
    type: object
  github_com_rabbytesoftware_quiver_core_internal_core_config.ArrowAutoRetry:
    properties:
      enabled:
        type: boolean
      retries:
        minimum: 0
        type: integer
    type: object
  github_com_rabbytesoftware_quiver_core_internal_core_config.Arrows:
    properties:
      auto_retry:
        $ref: '#/definitions/github_com_rabbytesoftware_quiver_core_internal_core_config.ArrowAutoRetry'
      extract_max_bytes:
        minimum: 1
        type: integer
      self_update_channel:
        type: string
      version_check_ttl:
        type: string
    type: object
  github_com_rabbytesoftware_quiver_core_internal_core_config.Auth:
    properties:
      pairing_code_ttl:
        type: string
      redeem_rate_limit:
        minimum: 1
        type: integer
      redeem_rate_window:
        type: string
    type: object
  github_com_rabbytesoftware_quiver_core_internal_core_config.Logger:
    properties:
      enabled:
        type: boolean
      level:
        type: string
    type: object
  github_com_rabbytesoftware_quiver_core_internal_core_config.Manifold:
    properties:
      fetch_timeout:
        type: string
      fletcher:
        $ref: '#/definitions/github_com_rabbytesoftware_quiver_core_internal_core_config.ManifoldFletcher'
    type: object
  github_com_rabbytesoftware_quiver_core_internal_core_config.ManifoldFletcher:
    properties:
      enabled:
        type: boolean
      min_stars:
        minimum: 0
        type: integer
      probe_limit:
        minimum: 1
        type: integer
    type: object
  github_com_rabbytesoftware_quiver_core_internal_core_config.Netbridge:
    properties:
      enabled:
        type: boolean
      ephemeral_port_end:
        maximum: 65535
        minimum: 1
        type: integer
      ephemeral_port_start:
        maximum: 65535
        minimum: 1
        type: integer
    type: object
  github_com_rabbytesoftware_quiver_core_internal_core_config.Search:
    properties:
      fetch_concurrency:
        minimum: 1
        type: integer
      per_provider_limit:
        minimum: 1
        type: integer
      provider_timeout:
        type: string
    type: object
  github_com_rabbytesoftware_quiver_core_internal_core_config.Vault:
    properties:
      index_ttl:
        type: string
      sweep_interval:
        type: string
      ttl:
        type: string
    type: object
...
host: localhost:40257
info:
  contact:
    name: Rabbyte Software
  description: Quiver is a decentralized package manager.
  title: Quiver API
  version: "0"
paths:
...
  /config:
    get:
      description: |-
        Returns the configuration three ways at once: running is what this process is using, configured is what the next start will use, and defaults is what ships in the binary.

        Every configuration change takes effect on restart, so running and configured differ whenever a change is still pending. restart_required names exactly those fields, so a client does not have to diff the documents itself.

        running carries no api section: the --host flag can override the configured host at start, so the daemon cannot report a bind address from configuration alone.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_rabbytesoftware_quiver_core_internal_api_libs.QueryResponse'
            - properties:
                data:
                  $ref: '#/definitions/github_com_rabbytesoftware_quiver_core_internal_api_v0_dto.ConfigDTO'
              type: object
        "500":
          description: The configuration file could not be read
          schema:
            $ref: '#/definitions/github_com_rabbytesoftware_quiver_core_internal_api_libs.ErrResponse'
      summary: Read the daemon configuration
      tags:
      - system
    patch:
      consumes:
      - application/json
      description: |-
        Applies a sparse configuration change. A field left out of the body is untouched, a field set to null is restored to its default, and a field carrying a value is set to it.

        Settings are accepted or refused one at a time: a body mixing valid and invalid settings persists the valid ones and reports the rest in rejected, including any key the daemon does not recognise. The response is 422 only when nothing at all could be applied.

        Nothing takes effect until the daemon restarts. Re-read GET /v0/config to see the change reflected in configured and listed in restart_required.
      parameters:
      - description: 'Sparse configuration document: any subset of the sections and
          settings returned by GET /v0/config, with null to restore a default'
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_rabbytesoftware_quiver_core_internal_app_usecases.Config'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_rabbytesoftware_quiver_core_internal_api_libs.QueryResponse'
            - properties:
                data:
                  $ref: '#/definitions/github_com_rabbytesoftware_quiver_core_internal_api_v0_dto.ConfigPatchResultDTO'
              type: object
        "400":
          description: Body could not be read
          schema:
            $ref: '#/definitions/github_com_rabbytesoftware_quiver_core_internal_api_libs.ErrResponse'
        "422":
          description: Every setting in the body was refused; rejected names each
            one and why
          schema:
            allOf:
            - $ref: '#/definitions/github_com_rabbytesoftware_quiver_core_internal_api_libs.QueryResponse'
            - properties:
                data:
                  $ref: '#/definitions/github_com_rabbytesoftware_quiver_core_internal_api_v0_dto.ConfigPatchResultDTO'
              type: object
      summary: Change the daemon configuration
      tags:
      - system
...
  /system/path:
    get:
      description: Reports whether ~/.quiver/bin exists, is on PATH, and, where the
        platform tracks it, whether it is configured to stay there (shell rc files
        on unix, the user Path on Windows).
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_rabbytesoftware_quiver_core_internal_api_libs.QueryResponse'
            - properties:
                data:
                  $ref: '#/definitions/github_com_rabbytesoftware_quiver_core_internal_api_v0_dto.PathStatusDTO'
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_rabbytesoftware_quiver_core_internal_api_libs.ErrResponse'
      summary: Read PATH status
      tags:
      - system
    post:
      description: 'Adds ~/.quiver/bin to PATH on explicit user action: appends a
        block to shell rc files on unix, sets the user Path on Windows. Idempotent
        — calling it again once already configured is a no-op.'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_rabbytesoftware_quiver_core_internal_api_libs.QueryResponse'
            - properties:
                data:
                  $ref: '#/definitions/github_com_rabbytesoftware_quiver_core_internal_api_v0_dto.PathStatusDTO'
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_rabbytesoftware_quiver_core_internal_api_libs.ErrResponse'
      summary: Set up PATH
      tags:
      - system
produces:
- application/json
swagger: "2.0"