# internal/api/v0/endpoints/system/handlers/handlers.go
package system

import (
	"io"
	"net/http"

	"github.com/gin-gonic/gin"

	"github.com/rabbytesoftware/quiver.core/internal/api/libs"
	"github.com/rabbytesoftware/quiver.core/internal/api/libs/apierr"
	apidto "github.com/rabbytesoftware/quiver.core/internal/api/v0/dto"
	"github.com/rabbytesoftware/quiver.core/internal/app/usecases"
)

// Handlers serves the daemon configuration endpoints.
type Handlers struct {
	svc     usecases.ConfigUsecase
	pathSvc usecases.PathUsecase
}

func New(
	svc usecases.ConfigUsecase,
	pathSvc usecases.PathUsecase,
) *Handlers {
	return &Handlers{svc: svc, pathSvc: pathSvc}
}

// Config returns the daemon configuration.
//
// @Summary      Read the daemon configuration
// @Description  Returns the configuration three ways at once: running is what this process is using, configured is what the next start will use, and defaults is what ships in the binary.
// @Description
// @Description  Every configuration change takes effect on restart, so running and configured differ whenever a change is still pending. restart_required names exactly those fields, so a client does not have to diff the documents itself.
// @Description
// @Description  running carries no api section: the --host flag can override the configured host at start, so the daemon cannot report a bind address from configuration alone.
// @Tags         system
// @Produce      json
// @Success      200  {object}  libs.QueryResponse{data=apidto.ConfigDTO}
// @Failure      500  {object}  libs.ErrResponse  "The configuration file could not be read"
// @Router       /config [get]
func (h *Handlers) Config(c *gin.Context) {
	view, err := h.svc.Get(c.Request.Context())
	if err != nil {
		status, msg := apierr.StatusAndMessage(err)
		libs.WriteErr(c, status, msg, "", err)
		return
	}

	libs.WriteQueryOK(c, apidto.ConfigDTOFrom(view))
}

// PatchConfig changes the daemon configuration.
//
// @Summary      Change the daemon configuration
// @Description  Applies a sparse configuration change. A field left out of the body is untouched, a field set to null is restored to its default, and a field carrying a value is set to it.
// @Description
// @Description  Settings are accepted or refused one at a time: a body mixing valid and invalid settings persists the valid ones and reports the rest in rejected, including any key the daemon does not recognise. The response is 422 only when nothing at all could be applied.
// @Description
// @Description  Nothing takes effect until the daemon restarts. Re-read GET /v0/config to see the change reflected in configured and listed in restart_required.
// @Tags         system
// @Accept       json
// @Produce      json
// @Param        body  body      usecases.Config  true  "Sparse configuration document: any subset of the sections and settings returned by GET /v0/config, with null to restore a default"
// @Success      200   {object}  libs.QueryResponse{data=apidto.ConfigPatchResultDTO}
// @Failure      400   {object}  libs.ErrResponse  "Body could not be read"
// @Failure      422   {object}  libs.QueryResponse{data=apidto.ConfigPatchResultDTO}  "Every setting in the body was refused; rejected names each one and why"
// @Router       /config [patch]
func (h *Handlers) PatchConfig(c *gin.Context) {
	body, err := io.ReadAll(c.Request.Body)
	if err != nil {
		libs.WriteErr(c, http.StatusBadRequest, "body could not be read", "")
		return
	}

	result, err := h.svc.Patch(c.Request.Context(), body)
	if err != nil {
		status, msg := apierr.StatusAndMessage(err)
		libs.WriteErr(c, status, msg, "", err)
		return
	}

	// A body whose settings were all refused reports every reason rather than
	// collapsing them into one message, so a client can mark each field.
	if len(result.Applied) == 0 && len(result.Rejected) > 0 {
		libs.WriteQueryWithStatus(
			c, http.StatusUnprocessableEntity, apidto.ConfigPatchResultDTOFrom(result),
		)
		return
	}

	libs.WriteQueryOK(c, apidto.ConfigPatchResultDTOFrom(result))
}

// @Summary      Read PATH status
// @Description  Reports whether ~/.quiver/bin exists, is on PATH, and, where the platform tracks it, whether it is configured to stay there (shell rc files on unix, the user Path on Windows).
// @Tags         system
// @Produce      json
// @Success      200  {object}  libs.QueryResponse{data=apidto.PathStatusDTO}
// @Failure      500  {object}  libs.ErrResponse
// @Router       /system/path [get]
func (h *Handlers) PathStatus(
	c *gin.Context,
) {
	st, err := h.pathSvc.Status(c.Request.Context())
	if err != nil {
		status, msg := apierr.StatusAndMessage(err)
		libs.WriteErr(c, status, msg, "", err)
		return
	}

	libs.WriteQueryOK(c, apidto.PathStatusDTOFrom(st))
}

// @Summary      Set up PATH
// @Description  Adds ~/.quiver/bin to PATH on explicit user action: appends a block to shell rc files on unix, sets the user Path on Windows. Idempotent — calling it again once already configured is a no-op.
// @Tags         system
// @Produce      json
// @Success      200  {object}  libs.QueryResponse{data=apidto.PathStatusDTO}
// @Failure      500  {object}  libs.ErrResponse
// @Router       /system/path [post]
func (h *Handlers) SetupPath(
	c *gin.Context,
) {
	st, err := h.pathSvc.Setup(c.Request.Context())
	if err != nil {
		status, msg := apierr.StatusAndMessage(err)
		libs.WriteErr(c, status, msg, "", err)
		return
	}

	libs.WriteQueryOK(c, apidto.PathStatusDTOFrom(st))
}